requirements.txt
python-socketio[asyncio_client]==4.6.1
python-engineio==3.14.2
structlog==25.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart
websockets>=12.0
aiortc>=1.5.0
av>=11.0.0
aiohttp>=3.9.0